Update patch set 8
Patch Set 8:

> i would have preferred if vpn interface for floating ips is also
 > handled here for dynamic vlans, instead of excluding. will give a
 > +1 now, with the request to address this in a follow up patch.

Ravindra could you just confirm Daya's request of taking this up in the future? We could also possibly note in the spec that this is out of scope for this current spec but will be handled in the future (Nitrogen?)

I rebased to resolve conflicts and sorted index.rst - waiting for your final confirmation before merging.

Patch-set: 8