Update patch set 6
Patch Set 6: Code-Review-1

(6 comments)

I started the review, but would really help to:

a. Clean up all the formatting changes from this review as they take up a huge chunk of the change
b. Do a dump-flows of the ingress/egress ACL in a scenario where there is a remote ACL on a single SG, and a scenario where there are two SGs and a remote ACL.
c. Run gate once the code is compiling

Patch-set: 6
Label: Code-Review=-1