Update patch set 14
Patch Set 14:

Alon, Im not so sure adding blueprint content via external documents will be a good way to proceed. Mainly because the external content is not controlled by git and gerrit code review. That is, I could modify the external content at any time and it doesnt have to pass code reviews.

This is for Sam, the PTL to decide.

As for your second comment, I will update that now, as there wont be resubmitting between tables on the ingress side, rather a resubmit from the Egress SFC classifier back to the Ingress side for the SFC SFF to process. Sam, Andre, and I discussed this and agreed this is a better approach.

Patch-set: 14
Label: Verified=0