Update patch set 17
Patch Set 17:

Subnet-Op-Data will become general purpose operational store for Subnets.   They will be populated with Dpn-To-Vpn list only for subnetRoute use-case.  

For all other use-case the Subnet-Op-Data will simply serve as an Operational Bucket for a Subnet until the subnet itself is used and removed by every VRF-driven-Flow.

As you can see from review link I posted earlier, subnet-op-data now has complete network-attributes that we can take advantage of.

This is similar to how we have vpn-interfaces Operational and vpn-instance-data Operational equivalents for vpn-interfaces config and vpn-instances config.

But for the patch to appear it will take couple more days.

I didn't want to block this review based on that one appearing.  But I was requesting if we can retain the elanInstances use until then instead of bringing in SubnetMap, was possible?

Vivek

Patch-set: 17
Label: Verified=0