Early implementation of classifier service
The idea of the classifier service is to replace the DataProcessor
approach. The rationale is that there is a huge amount of
heterogeneous events that may require adjustment of the global
operational state of the classifier. Having specific code to handle
each specific event would require a lot effort. The classifier service
works on this by having a generic processing that caters for any event
that may affect it's operational state.

The classifier service is based on the concept of ClassifierEntry, a tuple
that defines a classifier flow. A collection of this entries represents
the global state of the configuration and operational classifier, one
collection each. By comparing these two collections, we find the
entries that need to be added or removed from the operational state
based on the configuration state at any given time.

Change-Id: Icc33ac761297cea645a34f09d836d6364a3de0ca
Signed-off-by: Jaime Caamaño Ruiz <jaime.caamano.ruiz@ericsson.com>