Bug 8054: Fix IndexOutOfBoundsException in VrfEntryListener
Issue:
The exception occurs in case handling cleanupDpnForVpn in case of the
prefix not being an extra route as the usedRds is being tried to be
accessed to fetch the extraRoute(it is absent for the non-extra route
prefixes) and so throws the exception.

Fix:
A condition has been indroduced to fetch extraRoutes only if the prefix is
extra route else extraRoutes is set as absent.

Change-Id: Ie6ac4e4121e8dc1e418d21f77ccf41882b78d968
Signed-off-by: gobinath <gobinath@ericsson.com>