Added a new framework for Selecting centralized SNAT router.
Added a new HA framework for selecting the centralized SNAT router which
will use a weighted round robin to select the NAPT switch

A pseudo port will be added once the router is elected as the designated
switch and will be delete on router delete.

The switch selection/release happens in response to external router
create/delete or snat enable/disable

The flow programming will be done in response to   changes in
centralized switch model RouterToNaptSwitch.

Change-Id: I868b71f8a666925f72beddc811f4cc6b144bccc7
Signed-off-by: Aswin Suryanarayanan <asuryana@redhat.com>