Added support for multiple local_ips for DPN
This is part of the Policy based path selection for multiple VxLAN tunnels tasks.
Now when creating/updating a transportzone it will check for the local_ips information
from the ovsdb other_config entry - example:
local_ips="10.0.43.159:MPLS,11.11.11.11:DSL,ip:underlay-network"

If it is exist it will create the TZ based on it, if not it will take it
from the local_ip information.
It add support for multiple local ips on different underlay networks.

Change-Id: I3f7bba58e93372c1f48d0a23f844c32521011c03
Signed-off-by: yair <yair.zinger@hpe.com>