Update patch set 13
Patch Set 13: Code-Review-1

(1 comment)

Something else I noticed, but if it's part of the TODO we can leave it for later.

There are a number of "this needs Gerrit ..." comments; are these still pending? If they're not, the comments should be removed. If they are, is this patch safe to merge anyway?

Patch-set: 13
Label: Code-Review=-1