Update patch set 14
Patch Set 14:

Stephen,this patch is a work in progress, that's why I mention it's a first version in the comment. We need this patch in as a baseline so Jaime and I can split up the rest of the work. It's not production ready yet, we'll make it so in subsequent patches. Finally, it's ok the Todo comments about another patch.

Patch-set: 14
Reviewer: Gerrit User 2039 <2039@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0