Update patch set 2
Patch Set 2:

> what is the big difference between the patch sets - last one has 27
 > changes, but the other has 202?
 > 
 > This looks like something we want on Boron. Can you cherry-pick if
 > it applies?

Both patchset#1 and patchset#2 consist of 27 differences. I think @Alon rebased the patchset and there are some other changes in the file.

Yes @Sam, it would be helpful to have this patch in Stable/Boron. I tried to cherry-pick, but its not a clean-cherrypick. I'll try to manually cherrypick this patch after doing some validation. Thanks.

Patch-set: 2