Bug 8093:Fib entries for external Fixed Ips are not programmed
Problem Statement:
==================
In RouterDpnChangeLister, the extRouter parameter is passed as null and in
NatUtil.getExternalSubnetForRouterExternalIp() is called, router is passed
as null resulting in NullPointerException

Solution:
==========
If the router data is null, we need to get the routername from ConfigDS
and use that router name in NatUtil.getExternalSubnetForRouterExternalIp()
to avoid null pointer exception.

Note: Below code is added for external fixed IP information to be writing
into odl-fib entry for EVPN_RT5 use cases.

writeTx.submit()

Change-Id: I7acb7c52b045b6e37b500aef8a6bbe93a0e03777
Signed-off-by: karthikeyan <karthikeyan.k@altencalsoftlabs.com>