Bind Tenant provider VLAN port to L3-Pipeline
NetVirt currently missing binding of L3 services for provider
ports configured in Flat/Vlan private tenant networks.

A problems rises when trying to identify the rotuer instance that
we are receiving traffic from, as a private network can be connected
to multiple routers (when using multiple subnets in a network).

In order to identify the right router, we have added the creation of
VpnInstances for private networks (similar to what we do with external
networks). The VpnId of the private network is coded in metadata in
the dispatcher table(t17), before entering the L3 service.

In L3_GW_MAC_TABLE (t19) there is a match for a router_inteface mac
+ vpn-id of private network, then the metadata is overrided with
the vpn-id of the router.

Change-Id: I03a3e1d679333982eae9cafaa705415257f6c317
Signed-off-by: Tomer Pearl <tomer.pearl@hpe.com>