Add types of classifier entries
On the initial implementation a single ClassifierEntry would support
all the necessary information needed to render a classification for a
specific match on a specific node for a specific input interface. This
setup was proving difficult when suppressing from the dataplane
rendered items that would be shared between different classifications
like for example initialization flows or interface bindings.

With this patch the classifier entry now can be of different types
of smaller informational items that allow for a much more flexible
update southbound.

Change-Id: Ic355c0bf11d7c0c6f7264efb955d1f501bf905fd
Signed-off-by: Jaime Caamaño Ruiz <jaime.caamano.ruiz@ericsson.com>