Update patch set 7
Patch Set 7:

> Hi Tomer, I see some possible breakages for existing use cases
 > here..
 > 
 > Today, we allow creation for four types of networks – VLAN, VxLAN,
 > GRE and FLAT.
 > For all these types, we have a certain way of handling network
 > associations/dissociations to BGPVPNs.
 > 
 > As I see in this review, you’re creating VPNs per private VLAN/FLAT
 > network.
 > So are we saying that we don't support normal network based BGPVPN
 > forwarding for VLAN and flat networks? If yes, we still need to add
 > validations to error out appropriately.
 > Else, which I think is the case, this has to be reworked to handle
 > these types of networks similar to the way we handle routers today,
 > implemented with router-swapping kind of logic since:
 > 1. uninstallation of rules, dissociations etc need to be done
 > before such networks are associated to BGPVPNs.
 > 2. in case such networks are made a part of a router, similar
 > uninstallation of rules/flows and dissociations from existing VPN
 > needs to be done.
 > 
 > Let me know what you think..

Hi Abhinav,
So I have tried to figure this out from my side. Based on previous experience, when we tried to introduce external networks we added all kinds of checks to make sure we are in Flat/VLAN use-case. We are not familiar with BGPVPN use-case in Flat/VLAN tenant (private) networks.

If we are not sure that this indeed break something, and in order to meet the code freeze deadline, I think we should go forward with this and fix any bug in the upcoming bug fixes period. 

What do you think?

Patch-set: 7