BUG 8140 : Elect nodes for subnetroute on control path reboot
When a physical switch on the cloud management network in the cloud is
rebooted, it results in disconnection of datapath-nodes from the ODL
controller. Under such conditions, the subnetRoute is removed from all
the DPNs.
However, once the physical switch comes back up after reboot, the
datapath-nodes are all getting reconnected to the ODL controller.  During
this time in a scaled environment, the subnetRoute flow rules aren't
getting reapplied to the datapath-nodes.  This is because the election
process fails in figuring out a designated datapath-node for hosting
subnetroute, thereby the subnetRoute is not getting reconfigured on all
the datapath-nodes.

Change-Id: I4e2770a658e9d3b0c68fc52227845ed8546ecb97
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>