External networks subnetroute fix
Commit [1] changed the VpnSubnetRouteHandler APIs from using
yang based notifications to listening on subnetmap changes
from the datastore.

Subnetmap entities for subnets in Flat/VLAN external networks don't have a
vpn-name assigned to them, which caused a null pointer exception in
VpnSubnetRouteHandler after [1]

In this change, subnetmap is assinged a vpn-name, for subnets in
external Flat/VLAN networks, which is the subnet id itself.

[1] https://git.opendaylight.org/gerrit/#/c/53878/

Change-Id: Ic7593a5f59aa389f19729c5c5701fd9e1e02eea0
Signed-off-by: Tomer Pearl <tomer.pearl@hpe.com>