Fix erroneous space in DhcpAllocationPoolRemoveJob import statement
This shows up as "red ink" (error) in Eclipse as a Checkstyle violation.
I'm not sure why the CLI mvn Checkstyle does not fail forr this - it may
be a subtle difference between the exact version of Checkstyle
internally used by the Eclipse plugin and the one used by the Maven
build.

Change-Id: I3dacb8d5c6f6eb689bfadf034c2e1db7f30ce76f
Signed-off-by: Michael Vorburger <vorburger@redhat.com>