Update patch set 2
Patch Set 2: Code-Review-1

Thanks for the flow dumps. Still a bit confused - is this with a single compute node meaning both VMs are local to this OVS?
It would be easier to see if both VMs would use the same SG1 with remote_sg=SG1.

Looking at the flows you dumped though:

I thought that table 213 (or 243) needs to having matching on the src_ip (or dst_ip), and according to that set the metadata? For stateful this would be table 212 (or 242).
212 is the auto-learning table - it shouldn't have any non-learned rules.

In 214 (or 244), there used to be a match of the dst_ip (or src_ip) in addition to the metadata match - where are these now?

Patch-set: 2
Label: Code-Review=-1