Untangle bad class design: OpenFlow13Provider does not use DataBroker
It's generally best practice to not "reach over" through other services
to obtain injected services, but just let each class have (final) fields
with all of its dependant services.

In this particular case, the OpenFlow13Provider does not need to have a
DataBroker, so it's best to keep this simpler.  This makes writing unit
tests for OpenFlow13Provider more clear; for example, in
https://git.opendaylight.org/gerrit/#/c/54376/. 

Change-Id: I3924275c246ffca2da259a2f696a51282eab486a
Signed-off-by: Michael Vorburger <vorburger@redhat.com>