Bug 8001: Data validation failed for path CSIT TCs
Problem Description:
====================
On external network update or internet VPN update NAT code will trigger
the ExternalRoutersListener.update() to handle the updated changes
in NAT feature. During this invocation, for getting NAPT switch-id is
retrieved from the "router-to-napt-switch" DS. After retrieved the NAPT
switch-id from the DS, it contains the value as "0". There is no check for
NAPT switch-id validation.

clrRtsFromBgpAndDelFibTs() method is invoked with passing NAPT switch-id
as "0" value is throwing the data validation exceptions.

Solution:
===========
As part of this issue fix, invoked NAPT Switch selection if dpnId value is
"null" or "0" from "router-to-napt-switch" DS then go for electing new
Napt switch for existing router. This will resolve the data validation
exception. Since clrRtsFromBgpAndDelFibTs() method is never invoked with
NAPT switch-id as "0".

Change-Id: I851c4049397369871616834f59aff1eb16ce4472
Signed-off-by: karthikeyan <karthikeyan.k@altencalsoftlabs.com>