Update patch set 17
Patch Set 17:

(1 comment)

Hi Tomer,

There is no port-op data entries available for the external subnet, and as a result there is no designated node selected for the same.

All the ports if you happen to see on the external subnet are either router-gateway-ports (or) floating-ip-ports (or) dhcp-ports.  Such ports are not added to subnetmap itself and so as a result subnetrouteengine is unable to see those ports..

Did the PNF feature work after your commit, can you please point me to the link of the passing TC for the PNF ping?

see here for details on subnetmap for externalsubnet not having a single port and so the designation algorithm simply becomes a no-op:
https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-openstack-newton-nodl-v2-gate-stateful-carbon/778/robot/report/log.html#s1-s1-s3-t15-k5-k2-k1-k2-k18-k4

I tend to think that we should not block out this patch for PNF failure, because this patch is to get the basic subnetroute feature to work with connected nodes in the first place and can be re-used for PNF...

We can look further into PNF failure as a separate bug, as one more issue was found out by Koby (thanks to him) and he sent us a mail on the same.

We addressed all other comments from you.

Let me know what you think.

Vivek

Patch-set: 17