Bug 8170 - Loadbalancing nexthop group not updated after VM deletion
Issue:

When the last nexthop present in a DPN is deleted, the loadbalancing next
hop group is not updated. The group still contains the bucket pointing to
the deleted nexthop.

Fix:

There was an error while retrieving the endpointIpaddress for the DPN in
method getEndpointIpAddressForDPN. There was an error in converting the
IpAddress object to string. Appropriate conversion is used now.

Change-Id: I86cc77bd69446881d2ab669471f3142845e4d549
Signed-off-by: gobinath <gobinath@ericsson.com>