Bug 8170 - Loadbalancing nexthop group not updated after VM deletion
Issue:

When the last nexthop present in a DPN is deleted, the loadbalancing next
hop group is not updated. The group still contains the bucket pointing to
the deleted nexthop.

Fix:

The Vpntoextraroute and Prefixtointerface DS were not updated in case the
number of nexthops for an extraroute is greater than 1. Appropriate
condition has been added to handle this.

Change-Id: I86cc77bd69446881d2ab669471f3142845e4d549
Signed-off-by: gobinath <gobinath@ericsson.com>