Update patch set 3
Patch Set 3:

Adding N.Vivek for the review about l3vpn ipv6 internet access.

Some detailed information is available in the specification.
Kindly review the document.

I would also like to bring your attention to the fact that currently it is not possible to define a VPN as being an internet VPN ( default fallback processing will be done if more than one RT is caught ) or not.
I was thinking of adding a neutronvpn config parameter ( see below).
Kindly any answer is welcome.
+++ b/vpnservice/neutronvpn/neutronvpn-api/src/main/yang/neutronvpn.yang
@@ -133,7 +133,13 @@ module neutronvpn {
         leaf l3vni {
            type uint32;
         }
-    }
+
+        leaf isipv6vpninternet {
+           type boolean;
+           default false;
+           description "Boolean introducing the ability to configure a fallback mecanism for VPN importing that VPN";
+        }
+   }
 
     grouping network-attributes {
         leaf network-id {

Patch-set: 3
Label: Verified=0