bgpmanager: add afi_t parameter to add/delPrefix API and affFibentryToDS
An address family type is added to the bgpmanager API when advertising ,
removing a prefix. The modification also calls the thrift function calls
accordingly, as it has been modified previously in commit "Thrift
interface changes to support BGP VPNv6".
Also, in order to support FIB write for IPv4 and IPv6 prefixes, a new
parameter is added to addFibEntryToDS, in order to pass the afi
parameter.

Change-Id: I7b10831e8f9c480a848bd68b9c3516115851eaac
Signed-off-by: Noel De Prandieres <prandieres@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>