Update patch set 3
Patch Set 3:

As Koby mentioned the getSourceIpv4Address method needs work. The IPv4 desrializer is assuming the data handed to it is pointing to the IPv4 header, but the data includes the eth header also. Need to pass offset 12 or move the data pointer before calling deserialize.

These are DHCP packets in the exception. Notice the address in error: 134235392. That is 0x08004500 - which just happens to be 12 bytes into the dhcp packet (12 bytes is the dst and src mac) and then hit the eth header type for ipv4, 0x0800 and ip version and length: 0x45.

Patch-set: 3