Bug 8105: IllegalArgumentException in getSrcIp fix
1. reading IP and ARP payload was wrong, which is fixed by getting payload from
deserialised ethernet data based on IPv4 or ARP, getPayload will return
the right payload as the offsets are maintained in Packet based on
type of payload.
2. Tested these changes by creating 2 ports in neutron network, delete 1
port and send ping packets from deleted port to the other port.
First ARP packet hit and was able to retrieve IP address properly, and
then data traffic hit and was able to retrieve IP address.

Change-Id: I16af2e18e2c857a5529f4ffef1caf9da014d521f
Signed-off-by: Riyazahmed D Talikoti <riyazahmed.d.talikoti@ericsson.com>