Bug 8221 - Remove Wait Time To Install L3VPN Groups on OVS
Currently we have a wait of 1.5s to ensure L3VPN groups are programmed
correctly in OVS, before we start programming the L3VPN Flows. This needs
to be optimized further.
This patch makes use of the OpenFlowPlugin addGroup() RPC call to directly
install the group in the OVS. This will ensure that the subsequent installation
of flows will always find the required groups in the OVS.
Group installation via the FRM is still retained to ensure consistency in the
datastore.

Change-Id: Ie4c2cf2a6199bad2a27d067d444a42ae04d3f191
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>