Bug 8221 - Reduce Wait Time To Install L3VPN Groups on OVS
Currently we have a wait of 1.5s to ensure L3VPN groups are programmed
correctly in OVS, before we start programming the L3VPM Flows. This needs
to be optimized further.
This patch introduces direct RPC query to the OVS to check if the group
has been installed. If installed, we move ahead and start programming the
flows, Else we wait for 300ms and try to fetch the group information from
the OVS again. Retry count is 5.

Change-Id: Ie4c2cf2a6199bad2a27d067d444a42ae04d3f191
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>