Update patch set 2
Patch Set 2: Code-Review-1

Hi Gobinath,

I think this is to cover the bug that Koby reported on the fact that bucket number overlap happens for if macAddresses are learnt for discovered IPs on a subnet route.

If there is no bug id, can you please raise one and tag it to the commit message of this review.

Vivek

Patch-set: 2
Label: Code-Review=-1