Bug 8235 - Bug in group programming for directly connected subnet routes
Issue:

When an extra route is added(extra route with macaddress), the
loadbalancing group programmed for ecmp should contain either buckets
pointing to the local nexthop group or the tunnels. But in case of the
directly connected subnet routes, the groupIds of the loadbalancing group
and the local nexthop group turns out to be same, the latter overriding
it.

Analysis:

The loadbalancing groupId is computed from the prefix or the nexthop Ip
depending on whether the macadress is present or not(if mac is present,
the nexthopIp is used and prefix for otherwise). In case of the directly
connected subnet route, the macadress is present and so the nexthopIp
is used for computing the loadbalancing groupId. This overlaps with
local next hop group programmed for that nexthop.

Fix:

The issue arised in case the extra route contains macaddress(directly
connected subnet route) as the parameters used for computing the local
nexthop and the loadbalancing group turn out to be the same. Now
condition has been added so that no loadbalancing group is programmed in
case the subnet route is directly connected(ie macaddress is present). The
fib now directly points to the local nexthop group.

Change-Id: I6b0c5506f438ec1f816dfa0fdb045a8494950f48
Signed-off-by: gobinath <gobinath@ericsson.com>