Update patch set 8
Patch Set 8:

> > Hi Sam/Alon,
 > >
 > > The csit infra is playing out with issues here..
 > >
 > > Can we please review and merge this in?
 > >
 > > Vivek
 > 
 > Infra has some issues today, it would be good to see the logs after
 > PS8.
 > 
 > Assuming these "Unable to install group" errors will still happen
 > (since we already install the group via RPC), this is a very
 > problematic solution flooding the log with errors [1].
 > Can we think of a solution that does not contain these errors? This
 > is the reason I originally thought this utility belongs in
 > openflowplugin - allowing RPC + write to DS, resulting in one group
 > mod sent and no error.
 > 
 > [1] https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-newton-nodl-v2-gate-stateful-nitrogen/31/archives/odl1_karaf.log.gz

Then we need to enhance the OFplugin(not genius) to have a new method that will do both of them (store the group infor and send the RPC), but would make sure not to send the GroupMod again if the first invocation succeeded.  This kind of a patch may take time as it needs to go throu OFPlugin review cycles and I will discuss this with Hema here again.

Hi Alon,

BTW, we have swapped the calls in the latest PS (though that does not rule out GroupMod errors), but such GroupMod errors deosn't seem to get reported directly from the OFPlugin in the earlier CSIT runs if you have noticed.  All the 'Unable to install group' errors are in NextHopManager only triggered by previous PS..

Let me know what you think.

Patch-set: 8