Minor Checkstyle fix Custom Import Order error
This shows up as red in Eclipse, but the CLI mvn build missed this.
This is due to a more recent version of Checkstyle.

required for overall move to new Checkstyle version, see
https://git.opendaylight.org/gerrit/#/q/topic:bumpCheckstyle

also includes minor removal of (1) extra parenthesis, (2) generic <>
clean-up and (3) missing @Override from automatic clean-up which happens
on auto-save (technically unrelated to Checkstyle).

Conflicts:
	vpnservice/qosservice/impl/src/main/java/org/opendaylight/netvirt/qosservice/QosAlertManager.java

Change-Id: Ic650ffe9ac42cfbc7cad4c1c31c7553faf5fe603
Signed-off-by: Michael Vorburger <vorburger@redhat.com>
Signed-off-by: David Suarez <david.suarez.fuentes@ericsson.com>