Update patch set 5
Patch Set 5:

> (1 comment)

Why is that? My life is harder if I have to use a real databroker implemetation rather than a mock, it's just more work for me and I dont see the benefit.

I do agree on Future, shouldn't have mocked that and I don't know why I did now.

On the yang interfaces and Builders, it depends. On simple interfaces it's just easier to use a Builder. But when you have have to use multiple Builders to construct a complex structure it quickly becomes less work to mock (specially considering negative test cases) to a point where it outweights the illustration benefits of using the builders.

Patch-set: 5
Label: Verified=0