Update patch set 12
Patch Set 12: Code-Review+1

Hey, could you prioritize this commit in order to make inter-VPN link work again? This is the last commit for the whole refactor.

It must be cherry-picked to stable/carbon because the two previous commits where merged before the branchout and this makes the inter-VPN link fail in carbon also.

Patch-set: 12
Label: Code-Review=+1