BUG-8388: On egress, filter out SF bound packets
Classifier should not handle packets that are already classified and
bound towards the SFF. For this, we set C1 to 0xFFFFFF on the ACL flow.
Then on beguinning of egress pipeline, if C1 does not have that value
we assumed thats a packet bound for the SF and already classified so
it is resubmitteed to dispatcher. Otherwise, C1 is reset to 0 and
handling continues.

Change-Id: I1427a9bbec7fb0c8e36c241b727d39de450bc1ed
Signed-off-by: Jaime Caamaño Ruiz <jaime.caamano.ruiz@ericsson.com>