Update patch set 2
Patch Set 2:

> Seems wrong to be playing with the context headers to direct
 > packets through the pipeline. The whole idea of genius is to enable
 > the behavior.

The classifier is binding on egress for all interfaces as it has no way to know what the original destination of a packet might be. Genius does not provide a way to share information between ingress & egress. Internally it does REG6 to exchange some info. We could use a another REGX if that is preferred. Initially we think that it is less problematic to use a SFC specific field temporarily, that claiming ownership of a REG.

Patch-set: 2