Update patch set 3
Patch Set 3:

I prefer the way it is done in DhcpManager for DhcpInterfaceEventListener and DhcpInterfaceConfigListener. It would require not to use @Singleton for all these listeners, keep the check in single class and will work if we tomorrow support dynamic change of the flag. Just a preference though, would like inputs from others too on this.

Patch-set: 3