Bug 7939,8082: Table 19 flow nightmares
Table 19 flows rules aren't getting recycled appropriately
whenever VMs are moved into a routed subnet and then again
backed off.

The fix here attempts to make sure that Table 19 flow rules
are handled consistently when a subnet moves in/out of a router
and also when a router moves in/out of a bgpvpn.

Change-Id: Ief94b1b856a69c631052f11644db6a0b091f2aa7
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>