Update patch set 3
Patch Set 3: Code-Review+1

Talked to Vishal about this in the HWVTEP call today, and we agreed that this patch is OK to move forward. If we decide to support dynamic enable/disable of ODL DHCP, refactoring would be done then.

This patch needs to be backported to carbon and boron, so preference is to minimize change for backporting to these branches.

Patch-set: 3
Reviewer: Gerrit User 2029 <2029@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1