Bug 8476 Add support for logical switch replication mode
With OVS 2.6, the Logical Switch table, in the hardware_vtep schema, has
a new field, replication_mode. This bug will be used to add support to
netvirt, ovsdb, and genius for setting the replication_mode of logical
switches. The default will be for ODL to set the replication_mode
to "source_node" when creating logical switches.

Refer to [1] for more details on replication_mode.

[1] http://openvswitch.org/support/dist-docs/vtep.5.html

This patch depends on genius patch:
https://git.opendaylight.org/gerrit/#/c/57210/

Change-Id: I8581fa3bbc1af7c8d13bed2d66e351d020366ea3
Signed-off-by: Victor Pickard <vpickard@redhat.com>