Update patch set 9
Patch Set 9:

I could still see that Hairpinning rules in Table 19 are not getting cleaned up.  The clean up of Hairpinning rules is not related tot his review as the issue is happening before this review as well in upstream jobs:

https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-openstack-newton-nodl-v2-upstream-stateful-carbon/574/robot/report/log.html#s1-s4-t19

Other than that, could see that this review's gate run is same as the above upstream CSIT job.

Vivek

Patch-set: 9