Bug 8508 : Id-manager exception during releasing id for router
Description : Even though opendaylight-semantics is disabled,
the following EVPN NAT use-cases require VNI to be allocated from
OpenDaylight VNI Pool:

a. All traffic on external network will be using a VNI allocated for the
external-VPN aka Internet VPN.

b. Similarly all traffic from Non-NAPT switch to NAPTSwitch will be using
the VNI allocated for the router from the OpenDaylight VNI Pool.
For VNI-L3-L3 Fowarding feature also, a VNI must be allocated for every
router from the OpenDaylight VNI Pool as that VNI will be used for
Non-NAPT to NAPTSwitch traffic.

This fix makes that allocation as only release alone was done earlier.

Changes done to allocate a vni for router(when router-gatway-set done)
and been released during ext-router entry is deleted. In CSIT(having
non-EVPN use-case) were executed, an attempt to release the non-allocated
id was done during when ext-router entry is deleted resulting in exception.
Change done to make sure an id is allocated when router-gateway-set is done to
prevent such occurance of exception.

Change-Id: Ic1e696d6b4c8d0c56eac8b16ed0731d072b5b135
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>