Bug 8508 : Id-manager exception during releasing id for router
Description : Enough though opendaylight-semantics is disabled,
certain EVPN use-cases rerquire to carve-out VNI from the OpenDayLight
pool. Change done to allocate an vni for router(when router-gatwat-set)
and been released during ext-router entry is deleted. In CSIT(having
non-EVPN use-case) were executed, an attempt to release the non-allocated
id was done during when ext-router entry is deleted resulting in exception.
Change done to make sure an id is allocated when router-gateway-set is done to
prevent such occurance of exception.

Change-Id: Ic1e696d6b4c8d0c56eac8b16ed0731d072b5b135
Signed-off-by: cgowdru <chetan.arakere@altencalsoftlabs.com>