Bug 7939 : SubnetRoute functionality for bgpvpn broken.
IPs behind Neutron Interfaces are not discoverable for
router-associated-to-bgpvpn use-cases.

With this fix we are disallowing creation of subnetRoutes for internal
subnets off neutron routers.
More specifically, for just neutron routers, subnetRoutes won't be published.
However, if such subnets are part of a router that is within a bgpvpn, the
subnetroutes feature will continue to work.

For the dependent PNF review that requires subnetRoutes to be available,
the fix here needs to be further tweaked to make PNF discovery on internal
subnets of a VLAN-network to work.

As a result, we are making this review parent of the PNF review here:
https://git.opendaylight.org/gerrit/53984

Change-Id: If7a6b0e6ac26fed29551b73251efb4d672c0ceae
Signed-off-by: HANAMANTAGOUD V Kandagal <hanamantagoud.v.kandagal@ericsson.com>
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>