Bug 7939 : SubnetRoute functionality for bgpvpn broken.
IPs behind Neutron Interfaces are not discoverable for
router-associated-to-bgpvpn use-cases.

With this fix we are disallowing creation of subnetRoutes for internal
subnets off neutron routers.
More specifically, for just neutron routers, subnetRoutes won't be published.
However, if such subnets are part of a router that is within a bgpvpn, the
subnetroutes feature will continue to work.

For the PNF review that requires subnetRoutes to be available,
the fix here needs to be further tweaked to make PNF discovery on internal
subnets of a VLAN-network to work.

PNF Review referredin commit message here is:
https://git.opendaylight.org/gerrit/53984

Change-Id: I507aae2e3d1cbb72457f32b48e598ef0f6cda703
Signed-off-by: HANAMANTAGOUD V Kandagal <hanamantagoud.v.kandagal@ericsson.com>
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>
Signed-off-by: HANAMANTAGOUD V Kandagal <hanamantagoud.v.kandagal@ericsson.com>