natservice: update bgpvpn fallback pipeline rules for ipv6
When associating a BGPVPN to an external network, the identifier
associated to that BGPVPN is already used for all the underlying
external subnets declared in that network.
For IPv6 GUA subnets that are "behind that external network", that is
to say behind a router which has that that external network linked, that same
identifier is used to write a specific fallback route for those IPv6 GUA
to access the Internet through BGPVPN. Indeed, when the router has no
more entries in its FIB table, a fallback rule is written for IPv6 GUA
traffic against possibly BGPVPN IPv6 entries. That fallback rule
resubmits the packet to the L3 table, but with an ID set to the BGPVPN
id. Like that, all IPv6 traffic that has an entry associated may be
redirected to the correct DC-GW.
The modification here consists in adding that non prioritary fallback
rule, in the case a BGPVPN is present on external network, removing that
fallback rule, if the BGPVPN is detached from the external network.

Change-Id: Iebd754ae10fa6db08512980a98e07c3a02ebb6d5
Signed-off-by: Noel De Prandieres <prandieres@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>